pyproject.toml
[project]
name = "ship-tracker"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "discord.py>=2.4.0",
  "aiosqlite>=0.20.0",
  "python-dotenv>=1.0.1",
]

[tool.black]
line-length = 100
target-version = ["py310"]